* .gitignore file created; * CMakeLists.txt file created; * main.cpp improvements.

+++ Tasks/Task0.2/main.cpp
@@ -1,15 +1,14 @@
-#include <stdio.h>
 #include "broken_telephone.h"
-#include <stdlib.h>
-#include <string.h>
+#include <cstdio>
+#include <cstdlib>
+#include <cstring>
 
-int main()
-{
+int main() {
     char str[] = "FEFU is my home 1234567890";
-    int messageLen = strlen(str);
-    char *message = (char *)malloc(strlen(str)+1);
-    strcpy(message,str);
-    
+    int messageLen = (int) strlen(str);
+    char *message = (char *) malloc(strlen(str) + 1);
+    strcpy(message, str);
+
     message = lineNodeNoise1(message, messageLen);
     message = lineNodeNoise2(message, messageLen);
     message = lineNodeNoise3(message, messageLen);
